Patch auditwheel 3.2.0
This is the version of auditwheel included in the manylinux1 image
at the time of this commit. This patch is even more minimal than
previous patches, so should apply for much longer. What this patch
does is enforces that $ORIGIN is added to all RPATHs for so files
in the "<libname>.libs" directory (e.g. "htcondor.libs") that is
created by auditwheel to hold all of the external libraries.